load random country on app start
