+++ src/components/CountryContainer.js
@@ -1,4 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useMemo } from 'react';
+import defaultCountries from 'i18n-iso-countries/langs/en.json';
 import CountryDisplay from './CountryDisplay';
 import CountrySelector from './CountrySelector';
 
@@ -7,21 +8,40 @@ const BASE_URL = `http://localhost:${PORT}`;
 const FAVORITES_URL = `${BASE_URL}/favorites`;
 
 const CountryContainer = () => {
-    const [currCountry, setCurrCountry] = useState('Afghanistan');
-    // TODO: It would be cool to make currCountry start as a random country
+    const EXCLUDE = ["Czechia", "Micronesia, Federated States of", "Moldova, Rebublic of"];
+    // TODO: "India" is coming up as British Indian Ocean Territory...
+
+    const countryOptions = useMemo(() => {
+        const names = Object.values(defaultCountries.countries).map((value) =>
+            Array.isArray(value) ? value[1] : value
+        );
+        return names.filter((name) => name && !EXCLUDE.includes(name));
+    }, []);
+    // ^^ Better handle on the server side ?
+
+    const [currCountry, setCurrCountry] = useState(() => {
+        const randomIndex = Math.floor(Math.random() * countryOptions.length);
+        return countryOptions[randomIndex] || "Afghanistan";
+    });
     const [countryData, setCountryData] = useState(null);
     const [favsList, setFavsList] = useState([]);
-    
+    // TODO: Add vars for isLoading
+    // const [isDisplayLoaded, setIsDisplayLoaded] = useState(false);
+    // const [areFavoritesLoaded, setAreFavoritesLoaded] = useState(false);
+
     const headers = {'Content-Type': 'application/json'};
    
     const changeCurrCountry = (country) => setCurrCountry(country);
 
+    // TODO: combine fetchCountryInfo with the logic to create
+    // countryOptions and handle it all on the backend.
     const fetchCountryInfo = async (country) => {
         if (country === 'United States') country = 'United States of America';
         try {
             const res = await fetch(`https://restcountries.com/v2/name/${country}`);
             const message = `fetchCountryInfo: HTTP ${res.status}`;
             const data = await res.json();
+            console.log('countryData to be set:', data);
             setCountryData(data);
             console.log(message);
         } catch(err) {
@@ -40,9 +60,8 @@ const CountryContainer = () => {
             if (!res.ok) throw new Error(message);
             // DYK: The use of throw immediately moves the thread of execution to the catch block
             const data = await res.json();
-            console.log('Setting favorites:', data);
-            setFavsList(data);
             console.log('message:', message);
+            setFavsList(data);
         } catch (err) {
             console.error(err);
         }
@@ -63,6 +82,7 @@ const CountryContainer = () => {
             if (!res.ok) throw new Error(message);
             const data = await res.json();
             console.log(message, data);
+            fetchFavorites();
         } catch (err) {
             console.error(err);
         }
@@ -80,8 +100,8 @@ const CountryContainer = () => {
             const res = await fetch(FAVORITES_URL, deleteOptions);
             const message = `deleteFavorite: HTTP ${res.status}`;
             if (!res.ok) throw new Error(message);
-            // fetchFavorites();
             console.log(message);
+            fetchFavorites();
         } catch (err) {
             console.error(err);
         }
@@ -100,7 +120,7 @@ const CountryContainer = () => {
             const message = `toggleVisited: HTTP ${res.status}`;
             if (!res.ok) throw new Error(message);
             console.log(message);
-            // fetchFavorites();
+            fetchFavorites();
         } catch (err) {
             console.error(err);
         }
@@ -123,6 +143,7 @@ const CountryContainer = () => {
             />
             <CountrySelector
                 favsList={favsList}
+                countryOptions={countryOptions}
                 changeCurrCountry={changeCurrCountry}
                 fetchFavorites={fetchFavorites}
                 deleteFavorite={deleteFavorite}
