+++ src/index.js
@@ -1,8 +1,6 @@
-// entry point for application. Hangs React app off of #contents in index.html
-
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import { Provider } from 'react-redux';
+// import { Provider } from 'react-redux';
 import { store } from './store/index.js';
 import App from './components/App.js'
 
@@ -10,7 +8,9 @@ const el = document.getElementById('root');
 const root = createRoot(el);
 
 root.render(
-    <Provider store={store}>
+    // <Provider store={store}>
+    <React.StrictMode>
         <App />
-    </Provider>
+    </React.StrictMode>
+    // </Provider>
 );
